Switching over to using unittest instead of nosetest
